feat(company-hierarchy): B2B-3216 fetch company hierarchy info with storeconfig
