+++ apps/storefront/src/shared/service/b2b/graphql/users.ts
@@ -37,22 +37,6 @@ const checkCustomerBCEmail = (data: CustomFieldItems) => `{
   }
 }`;
 
-const companyHierarchyAndMasquerading = () => `
-query CompanyHierarchyAndMasquerading {
-	companySubsidiaries {
-		companyId
-		companyName
-		parentCompanyId
-		parentCompanyName
-		channelFlag
-	}
-	userMasqueradingCompany {
-		companyId
-		companyName
-		bcId
-	}
-}`;
-
 export interface UserEmailCheckResponse {
   data: {
     userEmailCheck: {
@@ -87,8 +71,3 @@ export const checkUserBCEmail = (data: CustomFieldItems) =>
     ...res.customerEmailCheck,
     isValid: res.customerEmailCheck.userType !== UserTypes.B2C,
   }));
-
-export const getCompanyHierarchyAndMasquerading = () =>
-  B3Request.graphqlB2B({
-    query: companyHierarchyAndMasquerading(),
-  });
