+++ apps/storefront/src/shared/service/b2b/index.ts
@@ -22,8 +22,8 @@ import {
   getCurrencies,
   getProductPricing,
   getStoreConfigsSwitchStatus,
-  getStorefrontConfig,
   getStorefrontConfigs,
+  getStorefrontConfigWithCompanyHierarchy,
   getStorefrontDefaultLanguages,
   getTaxZoneRates,
   getUserCompany,
@@ -164,6 +164,7 @@ export {
   getBcShoppingListDetails,
   getBCStoreChannelId,
   getCompanySubsidiaries,
+  getStorefrontConfigWithCompanyHierarchy,
   startUserMasqueradingCompany,
   endUserMasqueradingCompany,
   getVariantInfoBySkus,
@@ -172,7 +173,6 @@ export {
   getOrderStatusType,
   getQuoteExtraFieldsConfig,
   getShoppingListsCreatedByUser,
-  getStorefrontConfig,
   getStorefrontConfigs,
   getStoreConfigsSwitchStatus,
   getStorefrontDefaultLanguages,
